+++ example/spec/parallel/ingestGranule/IngestGranuleFailureSpec.js
@@ -70,15 +70,24 @@ describe('The Ingest Granule failure workflow', () => {
       inputPayload = await setupTestGranuleForIngest(config.bucket, inputPayloadJson, granuleRegex, testSuffix, testDataFolder);
       pdrFilename = inputPayload.pdr.name;
 
-      // add a non-existent file to input payload to cause lambda error
+      // add a file with invalid schema (missing path field), and a non-existent file to input payload.
+      // .cmr.json is for testing retrieving cmr information when granule is failed
       inputPayload.granules[0].files = [
         {
-          name: 'non-existent-file',
-          key: 'non-existent-path/non-existent-file',
+          key: 'no-path-field/no-path-field-file',
           bucket: config.bucket,
+          name: 'no-path-field-file',
         },
+        {
+          key: 'non-existent-path/non-existent-file.cmr.json',
+          bucket: config.bucket,
+          name: 'non-existent-file',
+          path: 'non-existent-path',
+        },
+        ...inputPayload.granules[0].files,
       ];
 
+      console.log(`testSuffix: ${testSuffix}, granuleId: ${inputPayload.granules[0].granuleId}`);
       workflowExecution = await buildAndExecuteWorkflow(
         config.stackName,
         config.bucket,
@@ -89,6 +98,7 @@ describe('The Ingest Granule failure workflow', () => {
       );
     } catch (error) {
       beforeAllFailed = true;
+      console.log('IngestGranuleFailure beforeAll caught error', error);
       throw error;
     }
   });
@@ -233,8 +243,16 @@ describe('The Ingest Granule failure workflow', () => {
       });
 
       expect(granule.status).toBe('failed');
-      expect(granule.error.Error).toBeDefined();
-      expect(granule.error.Cause).toBeDefined();
+      console.log('IngestGranuleFailure granule.error', granule.error);
+      const errors = JSON.parse(granule.error.errors || []);
+      expect(errors.length).toBeGreaterThanOrEqual(2);
+      errors.forEach((error) => {
+        const isSchemaValidationError = (error.Error === 'CumulusMessageAdapterExecutionError') &&
+          error.Cause.includes('jsonschema.exceptions.ValidationError');
+        const isPostgresWriteError = error.Error.includes('Failed writing files to PostgreSQL') &&
+          error.Cause.includes('null value in column "bucket" violates not-null constraint');
+        expect(isSchemaValidationError || isPostgresWriteError).toBeTrue();
+      });
     });
   });
 });